Added 10 supported op in the emulate switch, added new push and pop function for the stack

+++ src/main/scala/Emulator/Mapper.scala
@@ -16,7 +16,7 @@ class Mapper(mapper_type: Int) {
     0
   }
 
-  def write(address: Int, value: Int): Unit = {
+  def write(address: Int, value: Byte): Unit = {
 
   }
 }
